+++ packages/termunicode/CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.1.0
 
 - Initial version.
+
+## 0.2.0
+
+- Added new function `widthChars` that use the Characters package.
+- Updated `widthString` string to use `widthChars` function.